requirements.txt
Django==4.2.14
django-cors-headers==4.4.0
djangorestframework==3.15.2
matplotlib-inline==0.1.7
numpy==1.24.4
pandas==2.0.
